Added startup and receivers support
Project modules will now contain a startup.py. This file defines a single run
method that is called on Django startup.

In addition to startup support the project understands the receivers.py. This
file can be used to hook up signal receivers which otherwise end up in models.py
or other funky places where it should not be.